pyproject.toml
[tool.poetry]
name = "KikisBoilerplate"
version = "0.1.0"
description = ""
authors = ["Kiera Saltz"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
emoji = "^0.6.0"
python-dotenv = "^0.14.0"
disputils = "^0.1.3"
requests = "^2.24.0"
gitpython = "^3.1.8"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.782"

[tool.poetry.scripts]
bot = "bot:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
